+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-29 09:12:27
-// Commit: fix(core): prevent double tap duplicate action in core
+// Updated at: 2026-04-29 11:43:43
+// Commit: feat(core): integrate offline caching for core state
 
 class coreComponentState {
-  final String commitRef = '744';
-  final String lastMessage = 'fix(core): prevent double tap duplicate action in core';
-  final DateTime timestamp = DateTime.parse('2026-04-29T09:12:27+05:00');
+  final String commitRef = '753';
+  final String lastMessage = 'feat(core): integrate offline caching for core state';
+  final DateTime timestamp = DateTime.parse('2026-04-29T11:43:43+05:00');
 }